codes/requirements.txt
langchain_community
langchain_openai
streamlit
groq
faiss-cpu
python-dotenv
langchain_groq
bs4